Refactor convo form to fragment
