+++ app/src/main/java/com/lambdaschool/empoweredconversation/di/ConversationComponent.kt
@@ -1,13 +1,12 @@
 package com.lambdaschool.empoweredconversation.di
 
-import com.lambdaschool.empoweredconversation.activity.ConversationForm
-import com.lambdaschool.empoweredconversation.activity.MainActivity
+import com.lambdaschool.empoweredconversation.fragment.ConversationFragment
 import dagger.Component
 import javax.inject.Singleton
 
 @Singleton
 @Component(modules = [AppModule::class, ConversationRepoModule::class])
 interface ConversationComponent {
-    fun inject(conversationForm: ConversationForm)
+    fun inject(conversationFragment: ConversationFragment)
 
 }
