--- app/src/main/res/layout/activity_conversation_form.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto" android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            android:background="@color/colorPrimaryDark"
-    >
-
-        <androidx.cardview.widget.CardView android:layout_width="match_parent"
-                                           android:layout_height="wrap_content" android:layout_marginStart="10dp"
-                                           android:layout_marginEnd="10dp" app:cardCornerRadius="6dp"
-                                           app:cardElevation="5dp" app:cardUseCompatPadding="true">
-
-            <com.google.android.material.appbar.AppBarLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-            >
-
-                <androidx.appcompat.widget.Toolbar
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content" android:id="@+id/toolbar"
-                        app:titleTextColor="@color/colorPrimaryDark">
-
-                </androidx.appcompat.widget.Toolbar>
-
-            </com.google.android.material.appbar.AppBarLayout>
-        </androidx.cardview.widget.CardView>
-
-
-        <FrameLayout
-                android:id="@+id/drawer_layout"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-
-            <include layout="@layout/content_converstation_form"/>
-
-        </FrameLayout>
-
-    </LinearLayout>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
