+++ app/src/main/res/menu/nav_menu.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@+id/landingFragment" android:title="landingFragment"/>
+    <item android:id="@+id/landingFragment" android:title="Home"/>
+    <item android:id="@+id/conversationFragment" android:title="Use our Tool"/>
 </menu>
